Home 구조 변경
