+++ client/components/Data/homeData.js
@@ -1,0 +1,24 @@
+export const homeData = [
+  {
+    shopId: 3,
+    name: "가나",
+    cafeImageUrl: "https://cdn-icons-png.flaticon.com/128/3563/3563996.png",
+    stamp: 0,
+    menu: ["스콘", "아이스아메리카노"],
+    menuImageUrl: [
+      "https://cdn.paris.spl.li/wp-content/uploads/2023/02/%EC%8D%B8%EB%84%A4%EC%9D%BC%EC%86%8C%EA%B8%88%EB%B2%84%ED%84%B0%EC%8A%A4%EC%BD%98_02.jpg",
+      "123",
+    ],
+  },
+  {
+    shopId: 4,
+    name: "가나",
+    cafeImageUrl: "https://cdn-icons-png.flaticon.com/128/3563/3563996.png",
+    stamp: 0,
+    menu: ["스콘", "아이스아메리카노"],
+    menuImageUrl: [
+      "https://cdn.paris.spl.li/wp-content/uploads/2023/02/%EC%8D%B8%EB%84%A4%EC%9D%BC%EC%86%8C%EA%B8%88%EB%B2%84%ED%84%B0%EC%8A%A4%EC%BD%98_02.jpg",
+      "123",
+    ],
+  },
+];
